+++ packages/core/src/jsx-runtime/get-style.ts
@@ -4,6 +4,7 @@ import CoreError from '../utils/error';
 export const CSS_LIST = {
   backgroundColor: ['*'],
   border: ['*'],
+  boxSizing: ['*'],
   borderBottom: ['*'],
   borderBottomColor: ['*'],
   borderBottomStyle: ['*'],
@@ -105,4 +106,25 @@ function checkStyleValue(
       });
     }
   }
+}
+
+export function getStyleFromString(styleString: string, key: string) {
+  if (typeof styleString !== 'string') {
+    return undefined;
+  }
+
+  const styles = styleString.split(';').filter((s) => s);
+
+  for (const style of styles) {
+    const [keyStyle, value] = style.split(':');
+
+    if (key === keyStyle) {
+      return {
+        key: keyStyle,
+        value,
+      };
+    }
+  }
+
+  return undefined;
 }